requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.0
transformers==4.35.0
gunicorn==21.2.0
huggingface_hub==0.19.4
drain3==0.9.11
python-dotenv==1.0.0